Admin Auto Approve, Next PayDate, Table Order Desc

+++ application/views/super_admin/inc/product_list_content.php
@@ -63,9 +63,15 @@
                                 <td><?= $row->latest_price ?></td>
                                 <td><?= $row->sell_price ?></td>
                                 <td><?= $row->all_time_sell ?></td>
-                                <td><a onclick="return confirm('Want to delete?');" href="<?= base_url() ?>super_admin/deleteproductbyid/<?= $row->pro_id ?>" class="btn btn-secondary btn-block mt-0 tooltips" data-placement="top" data-toggle="tooltip" data-original-title="Delete">
-                                        <i class="fas fa-times"></i>
-                                    </a></td>
+                                <?php
+                                if ($row->total_stock == NULL) { ?>
+                                    <td><a onclick="return confirm('Want to delete?');" href="<?= base_url() ?>super_admin/deleteproductbyid/<?= $row->pro_id ?>" class="btn btn-secondary btn-block mt-0 tooltips" data-placement="top" data-toggle="tooltip" data-original-title="Delete">
+                                            <i class="fas fa-times"></i>
+                                        </a></td>
+                                <?php } else {
+                                    echo "<td><strong>N/A</strong></td>";
+                                }
+                                ?>
                             </tr>
                         <?php endforeach; ?>
                     </tbody>
